changed pair counter to be uint64_t

+++ malis/malis_cpp.cpp
@@ -5,6 +5,7 @@
 #include <vector>
 #include <queue>
 #include <map>
+#include <stdint.h>
 using namespace std;
 
 template <class T>
@@ -29,7 +30,7 @@ class AffinityGraphCompare{
 void malis_loss_weights_cpp(const int nVert, const int* seg,
                const int nEdge, const int* node1, const int* node2, const float* edgeWeight,
                const int pos,
-               int* nPairPerEdge){
+               uint64_t* nPairPerEdge){
 
 
     /* Disjoint sets and sparse overlap vectors */
